+++ src/packages/@app/repositories/guide.repository.ts
@@ -19,6 +19,12 @@ export class GuideRepository {
     return document;
   };
 
+  /**
+   * Finds a single document that matches the query.
+   *
+   * @param {Object} query - The query object used to find the document.
+   * @returns {Promise<Object>} - A promise that resolves to the found document.
+   */
   public findOne = async (query) => {
     return await RxDB.getInstance()
       .rxdb.guide.findOne({
@@ -26,6 +32,12 @@ export class GuideRepository {
       })
       .exec();
   };
+
+  /**
+   * @param {Object} query - The query object used to find the document.
+   * @param {Object} updateData - The data to update the found document with.
+   * @returns {Promise<void>} - A promise that resolves when the update is complete.
+   */
   public update = async (query, updateData) => {
     const rxCollection = RxDB.getInstance().rxdb.guide;
     const foundDocs = await rxCollection
