+++ src/packages/@deprecate/components/environments/enviroments-panel/EnvironmentPanel.svelte
@@ -20,7 +20,6 @@
     workspaceLevelPermissions,
   } from "$lib/utils/constants/permissions.constant";
   import Tooltip from "@library/ui/tooltip/Tooltip.svelte";
-    import WelcomeBox from "@library/ui/popover/Popover.svelte";
 
   export let environmentRepositoryMethods: EnvironmentRepositoryMethods;
   export let environmentServiceMethods: EnvironmentServiceMethods;
@@ -187,9 +186,6 @@
     {/if}
   </div>
 {/if}
-<div>
-  <WelcomeBox/>
-</div>
 
 <style lang="scss">
   .env-panel {
