+++ src/packages/@library/ui/carousel/index.ts
@@ -1,0 +1,1 @@
+export { default as Carousel } from "../carousel/Carousel.svelte";
