+++ src/packages/@library/ui/popover/index.ts
@@ -1,2 +1,1 @@
 export { default as Popover } from "./Popover.svelte";
-export {default as Carousel}  from "./Carousel.svelte";
